validateEventTypesFilters
