+++ src/back/routes/events.ts
@@ -2,11 +2,9 @@ import { WithAuth, auth } from 'decentraland-gatsby/dist/entities/Auth/middlewar
 import handleAPI from 'decentraland-gatsby/dist/entities/Route/handle'
 import routes from 'decentraland-gatsby/dist/entities/Route/routes'
 import { Request } from 'express'
-import isArray from 'lodash/isArray'
 
-import { EventFilterSchema } from '../../shared/types/events'
 import { EventsService } from '../services/events'
-import { validateProposalId, validateRequiredString } from '../utils/validations'
+import { validateEventTypesFilters, validateProposalId, validateRequiredString } from '../utils/validations'
 
 import { discourseComment } from './webhooks'
 
@@ -18,16 +16,8 @@ export default routes((route) => {
 })
 
 async function getLatestEvents(req: Request) {
-  const { event_type } = req.query
-  const filters: Record<string, unknown> = {}
-  if (event_type) {
-    filters.event_type = isArray(event_type) ? event_type : [event_type]
-  }
-  const parsedEventTypes = EventFilterSchema.safeParse(filters)
-  if (!parsedEventTypes.success) {
-    throw new Error('Invalid event types: ' + parsedEventTypes.error.message)
-  }
-  return await EventsService.getLatest(parsedEventTypes.data)
+  const eventTypes = validateEventTypesFilters(req)
+  return await EventsService.getLatest(eventTypes)
 }
 
 async function voted(req: WithAuth) {
